pyproject.toml
[project]
name = "ai-agent15-status-service"
version = "2.0.0"
description = "Status Service + Notification Service for HRMIS Recruitment Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "mysql-connector-python>=9.4.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "redis>=5.0.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
    "python-consul>=1.1.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.4.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "Pillow>=10.3.0",
    "python-multipart>=0.0.9",
    "pywebpush>=2.0.0",
    "cryptography>=42.0.0",
    # AI chatbot
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "google-generativeai>=0.8.0",
    "croniter>=2.0.5",
    "matplotlib>=3.9.0",
    "reportlab>=4.2.0",
    "mcp>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "moto[s3]>=5.0.0",
    "freezegun>=1.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
